Move comparison and diffing code to a new module.
This includes various minor refactorings to try to clean things up
and provide better error messages.
