+++ tests/testsuite/config.rs
@@ -5,7 +5,8 @@ use cargo::util::config::{self, Config, SslVersionConfig, StringList};
 use cargo::util::interning::InternedString;
 use cargo::util::toml::{self, VecStringOrBool as VSOB};
 use cargo::CargoResult;
-use cargo_test_support::{normalized_lines_match, paths, project, t};
+use cargo_test_support::compare::normalized_lines_match;
+use cargo_test_support::{paths, project, t};
 use serde::Deserialize;
 use std::borrow::Borrow;
 use std::collections::{BTreeMap, HashMap};
