+++ tests/testsuite/lockfile_compat.rs
@@ -1,8 +1,9 @@
 //! Tests for supporting older versions of the Cargo.lock file format.
 
+use cargo_test_support::compare::lines_match;
 use cargo_test_support::git;
 use cargo_test_support::registry::Package;
-use cargo_test_support::{basic_lib_manifest, basic_manifest, lines_match, project};
+use cargo_test_support::{basic_lib_manifest, basic_manifest, project};
 
 #[cargo_test]
 fn oldest_lockfile_still_works() {
